Movement set to only in a box
Managed to make the boids only move inside a box but is currently only swaying side to side, think it has to be because the movement to avoid being outside the box is too  strong. This will need some tweaking. Would also like to implement some more fun movement like a  swirl to give it a pretty movement.